# public/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Inventory & Order Management API",
    "version": "1.0.0",
    "description": "REST API for authentication, catalog, contacts, inventory, orders, dashboard, and reports. Login first, copy data.token, and authorize with the Bearer token."
  },
  "servers": [{ "url": "http://localhost:8088", "description": "Laravel API" }],
...
  "paths": {
...
    "/api/reports/monthly-sales": { "get": { "tags": ["Reports"], "summary": "Get delivered sales grouped by month", "security": [{ "bearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/Year" }], "responses": { "200": { "$ref": "#/components/responses/Success" } } } },
...
    "/reports/top-products": { "get": { "tags": ["Go Reports"], "summary": "Get top products from Go", "servers": [{ "url": "http://localhost:8090" }], "parameters": [{ "$ref": "#/components/parameters/Limit" }], "responses": { "200": { "$ref": "#/components/responses/Success" } } } },
    "/reports/monthly-sales": { "get": { "tags": ["Go Reports"], "summary": "Get monthly sales from Go", "servers": [{ "url": "http://localhost:8090" }], "parameters": [{ "$ref": "#/components/parameters/Year" }], "responses": { "200": { "$ref": "#/components/responses/Success" } } } }
  },
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "Sanctum token" } },
    "parameters": {
...
      "Limit": { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 10 } },
      "Year": { "name": "year", "in": "query", "schema": { "type": "integer", "minimum": 2000, "example": 2026 } }
    },
    "schemas": {
...
      "ApiResponse": { "type": "object", "properties": { "success": { "type": "boolean" }, "message": { "type": "string" }, "data": {} } },
...
    },
...
    "responses": {
      "Success": { "description": "Successful response", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" } } } },
...
    }
  }
}